Add sqldelight to shared library

+++ AndroidStudioProjects/HelloWorld/hello_world/src/commonMain/kotlin/com/example/helloworld/DataModels/PlaceholderResult.kt
@@ -3,4 +3,4 @@ package com.example.helloworld.DataModels
 import kotlinx.serialization.Serializable
 
 @Serializable
-class PlaceholderResult(val id: Int, val userId: Int, val title: String, val completed: Boolean)
+class PlaceholderResult(val id: Long, val userId: Long, val title: String, val completed: Boolean)
